+++ pcsx2/VU.h
@@ -128,7 +128,7 @@ struct alignas(16) VURegs
 
 	// flags/cycle are needed by VIF dma code, so they have to be here (for now)
 	// We may replace these by accessors in the future, if merited.
-	u32 cycle;
+	u64 cycle;
 	u32 flags;
 
 	// Current opcode being interpreted or recompiled (this var is used by Interps
@@ -156,7 +156,7 @@ struct alignas(16) VURegs
 	u32 statusflag;
 	u32 clipflag;
 
-	s32 nextBlockCycles;
+	s64 nextBlockCycles;
 
 	u8* Mem;
 	u8* Micro;
@@ -164,7 +164,7 @@ struct alignas(16) VURegs
 	u32 xgkickaddr;
 	u32 xgkickdiff;
 	u32 xgkicksizeremaining;
-	u32 xgkicklastcycle;
+	u64 xgkicklastcycle;
 	u32 xgkickcyclecount;
 	u32 xgkickenable;
 	u32 xgkickendpacket;
